+++ src/main/java/com/gestorinventario/entity/MarcaEntity.java
@@ -14,7 +14,7 @@ public class MarcaEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_marca")
-    private int idMarca;
+    private Long idMarca;
 
     @Column(name = "nombre_marca")
     private String nombreMarca;
