+++ src/main/java/com/gestorinventario/repository/AlmacenamientoRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.AlmacenamientoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlmacenamientoRepository extends JpaRepository<AlmacenamientoEntity, Integer> {
+public interface AlmacenamientoRepository extends JpaRepository<AlmacenamientoEntity, Long> {
 }
