+++ src/main/java/com/gestorinventario/repository/AreasRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.AreasEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AreasRepository extends JpaRepository<AreasEntity, Integer> {
+public interface AreasRepository extends JpaRepository<AreasEntity, Long> {
 }
