+++ src/main/java/com/gestorinventario/repository/EquipoRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.EquipoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipoRepository extends JpaRepository<EquipoEntity, Integer> {
+public interface EquipoRepository extends JpaRepository<EquipoEntity, Long> {
 }
