+++ src/main/java/com/gestorinventario/repository/HistoricoAsignacionRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.HistoricoAsignacionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoricoAsignacionRepository extends JpaRepository<HistoricoAsignacionEntity, Integer> {
+public interface HistoricoAsignacionRepository extends JpaRepository<HistoricoAsignacionEntity, Long> {
 }
