+++ src/main/java/com/gestorinventario/repository/HistoricoEquiposRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.HistoricoEquiposEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoricoEquiposRepository extends JpaRepository<HistoricoEquiposEntity, Integer> {
+public interface HistoricoEquiposRepository extends JpaRepository<HistoricoEquiposEntity, Long> {
 }
