+++ src/main/java/com/gestorinventario/repository/MarcaRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.MarcaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarcaRepository extends JpaRepository<MarcaEntity, Integer> {
+public interface MarcaRepository extends JpaRepository<MarcaEntity, Long> {
 }
