+++ src/main/java/com/gestorinventario/repository/MemoriaRamRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.MemoriaRamEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemoriaRamRepository extends JpaRepository<MemoriaRamEntity, Integer> {
+public interface MemoriaRamRepository extends JpaRepository<MemoriaRamEntity, Long> {
 }
