+++ src/main/java/com/gestorinventario/repository/ProcesadorRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.ProcesadorEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProcesadorRepository extends JpaRepository<ProcesadorEntity, Integer> {
+public interface ProcesadorRepository extends JpaRepository<ProcesadorEntity, Long> {
 }
