+++ src/main/java/com/gestorinventario/repository/TipoAlmacenamientoRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.TipoAlmacenamientoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoAlmacenamientoRepository extends JpaRepository<TipoAlmacenamientoEntity, Integer> {
+public interface TipoAlmacenamientoRepository extends JpaRepository<TipoAlmacenamientoEntity, Long> {
 }
