+++ src/main/java/com/gestorinventario/repository/UsuarioAsignadoRepository.java
@@ -3,5 +3,5 @@ package com.gestorinventario.repository;
 import com.gestorinventario.entity.UsuarioAsignadoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioAsignadoRepository extends JpaRepository<UsuarioAsignadoEntity, Integer> {
+public interface UsuarioAsignadoRepository extends JpaRepository<UsuarioAsignadoEntity, Long> {
 }
